move Errors class from file.py to errors.py
